requirements.txt
streamlit
groq
python-dotenv
httpx
pypdf
sentence-transformers
faiss-cpu
numpy